refactor: Change file structure
