+++ src/actions/index.test.js
@@ -8,12 +8,12 @@ import {
   setCategories,
   loadRestaurants,
   setRestaurants,
-} from './actions';
+} from '.';
 
 const middlewares = [thunk];
 const mockStore = configureStore(middlewares);
 
-jest.mock('./services/api');
+jest.mock('../services/api');
 
 describe('actions', () => {
   let store;
