+++ src/container/CategoriesContainer/index.test.jsx
@@ -2,9 +2,9 @@ import { render, fireEvent } from '@testing-library/react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import CategoriesContainer from './CategoriesContainer';
+import CategoriesContainer from '.';
 
-import CATEGORIES from '../fixtures/categories';
+import CATEGORIES from '../../../fixtures/categories';
 
 describe('CategoriesContainer', () => {
   const dispatch = jest.fn();
