+++ src/container/RegionsContainer/index.jsx
@@ -3,9 +3,9 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectRegion,
   loadRestaurants,
-} from './actions';
+} from '../../actions';
 
-import { get } from './utils';
+import { get } from '../../utils';
 
 export default function RegionsContainer() {
   const dispatch = useDispatch();
