+++ src/container/RestaurantsContainer/index.test.jsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react';
 
 import { useSelector } from 'react-redux';
 
-import RestaurantsContainer from './RestaurantsContainer';
+import RestaurantsContainer from '.';
 
 test('RestaurantsContainer', () => {
   useSelector.mockImplementation((selector) => selector({
