+++ src/utils/index.test.js
@@ -1,4 +1,4 @@
-import { get, equal } from './utils';
+import { get, equal } from '.';
 
 test('get', () => {
   const state = {
